✨ Use smooth curves instead of arcs
