+++ Sources/DynamicNotchKit/Views/NotchShape.swift
@@ -1,0 +1,120 @@
+//
+//  NotchShape.swift
+//  DynamicNotchKit
+//
+//  Created by Kai Azim on 2023-08-24.
+//
+
+import SwiftUI
+
+struct NotchShape: Shape {
+    var topCornerRadius: CGFloat {
+        bottomCornerRadius - 5
+    }
+
+    var bottomCornerRadius: CGFloat
+
+    init(cornerRadius: CGFloat? = nil) {
+        if cornerRadius == nil {
+            self.bottomCornerRadius = 11
+        } else {
+            self.bottomCornerRadius = cornerRadius!
+        }
+    }
+
+    var animatableData: CGFloat {
+        get { bottomCornerRadius }
+        set { bottomCornerRadius = newValue }
+    }
+
+    func path(in rect: CGRect) -> Path {
+        var path = Path()
+
+        path.move(
+            to: CGPoint(
+                x: rect.minX,
+                y: rect.minY
+            )
+        )
+
+        path.addQuadCurve(
+            to: CGPoint(
+                x: rect.minX + topCornerRadius,
+                y: rect.minY + topCornerRadius
+            ),
+            control: CGPoint(
+                x: rect.minX + topCornerRadius,
+                y: rect.minY
+            )
+        )
+
+        path.addLine(
+            to: CGPoint(
+                x: rect.minX + topCornerRadius,
+                y: rect.maxY - bottomCornerRadius
+            )
+        )
+
+        path.addQuadCurve(
+            to:  CGPoint(
+                x: rect.minX + topCornerRadius + bottomCornerRadius,
+                y: rect.maxY
+            ),
+            control: CGPoint(
+                x: rect.minX + topCornerRadius,
+                y: rect.maxY
+            )
+        )
+
+        path.addLine(
+            to: CGPoint(
+                x: rect.maxX - topCornerRadius - bottomCornerRadius,
+                y: rect.maxY
+            )
+        )
+
+        path.addQuadCurve(
+            to: CGPoint(
+                x: rect.maxX - topCornerRadius,
+                y: rect.maxY - bottomCornerRadius
+            ),
+            control: CGPoint(
+                x: rect.maxX - topCornerRadius,
+                y: rect.maxY
+            )
+        )
+
+        path.addLine(
+            to: CGPoint(
+                x: rect.maxX - topCornerRadius,
+                y: rect.minY + bottomCornerRadius
+            )
+        )
+
+        path.addQuadCurve(
+            to: CGPoint(
+                x: rect.maxX,
+                y: rect.minY
+            ),
+            control: CGPoint(
+                x: rect.maxX - topCornerRadius,
+                y: rect.minY
+            )
+        )
+
+        path.addLine(
+            to: CGPoint(
+                x: rect.minX,
+                y: rect.minY
+            )
+        )
+
+        return path
+    }
+}
+
+#Preview {
+    NotchShape()
+        .frame(width: 200, height: 32)
+        .padding(10)
+}
